当用户已经创建过房间，再次创建不会创建新房间会返回以前创建的房间ID

+++ Server/Model/ExtensionCode/Module/Game/RoomManageComponent.cs
@@ -17,7 +17,7 @@ namespace ETModel
     {
         public class RoomModel
         {
-            public long RoomId;
+            public int RoomId;
             
             public long PlayerId;
